Fixing bind address issue
The pkg/server/server.go code uses the config variable named address
to determin the bind address, but v0.4.0 code had --bind as the command
line parameter and it was saved into the bind variable in the code. This
means that in v0.4.0 you are unable to bind to a different interface.
This brings the v0.4.0 code up to match the master branch and correctly
name the command line parameter address to match the rest of the code.

Fixes https://github.com/kubernetes-sigs/aws-iam-authenticator/issues/248